### 2021-07-19 #3514 @whedon
[PRE REVIEW]: Symfem: a symbolic finite element definition library
**Submitting author:** @mscroggs (<a href="http://orcid.org/0000-0002-4658-2443">Matthew Scroggs</a>)
**Repository:** <a href="https://github.com/mscroggs/symfem" target ="_blank">https://github.com/mscroggs/symfem</a>
**Version:** v2021.7.6
