### 2021-07-20 #3514 @Nikoleta-v3
👋 @felipefr & @oalii & @sigvaldm & @iammix  would any of you be willing to review this submission for JOSS?

The submission I'd like you to review is titled: "Symfem: a symbolic finite element definition library". You can find more information at the top of this Github issue 🆙
