### 2021-07-20 ed3c02c
document python scripts in root folder
