+++ ++version.py
@@ -1,3 +1,9 @@
+"""
+This script increases the version number of Symfem in all the relevant
+locations. Once this has been run and the code pushed, Symfembot will
+automatically create a new version tag on GitHub.
+"""
+
 import json
 from datetime import datetime
 
