### 2021-07-20 #3514 @felipefr
Hello @Nikoleta-v3 <https://github.com/Nikoleta-v3>,

Yes I can, but I can only return mine in one month (maybe a little less).
What are the deadlines?

Felipe Rocha @felipefr

On Tue, 20 Jul 2021 at 13:00, Konstantinos Mixios ***@***.***>
wrote:

> Hello @Nikoleta-v3 <https://github.com/Nikoleta-v3>,
>
> yes I could. Unless there is a time limit and the review should be
> completed by the end of this month.
>
