### 2021-07-20 #3514 @Nikoleta-v3
Thank you both for getting back to me! 🙏🏻

There is no hard deadline. We aim for reviews to be completed within about 2-4 weeks. I can use Whedon (our bot) to set an automatic reminder in a month if you are both still happy to do the review 😄

I understand that this is a busy period and understand if you are unable to review at the present time.
