+++ symfem/vectors.py
@@ -88,3 +88,40 @@ def point_in_triangle(point, triangle):
     v = (dot00 * dot12 - dot01 * dot02) / det
 
     return u >= 0 and v >= 0 and u + v <= 1
+
+
+def point_in_tetrahedron(point, tetrahedron):
+    """Check if a point is inside a tetrahedron."""
+    v0 = vsub(tetrahedron[3], tetrahedron[0])
+    v1 = vsub(tetrahedron[2], tetrahedron[0])
+    v2 = vsub(tetrahedron[1], tetrahedron[0])
+    v3 = vsub(point, tetrahedron[0])
+
+    dot00 = vdot(v0, v0)
+    dot01 = vdot(v0, v1)
+    dot02 = vdot(v0, v2)
+    dot03 = vdot(v0, v3)
+    dot11 = vdot(v1, v1)
+    dot12 = vdot(v1, v2)
+    dot13 = vdot(v1, v3)
+    dot22 = vdot(v2, v2)
+    dot23 = vdot(v2, v3)
+
+    det = dot00 * (dot11 * dot22 - dot12 * dot12)
+    det += dot01 * (dot02 * dot12 - dot01 * dot22)
+    det += dot02 * (dot01 * dot12 - dot11 * dot02)
+
+    u = (dot11 * dot22 - dot12 * dot12) * dot03
+    u += (dot02 * dot12 - dot01 * dot22) * dot13
+    u += (dot01 * dot12 - dot02 * dot11) * dot23
+    u /= det
+    v = (dot12 * dot02 - dot01 * dot22) * dot03
+    v += (dot00 * dot22 - dot02 * dot02) * dot13
+    v += (dot02 * dot01 - dot00 * dot12) * dot23
+    v /= det
+    w = (dot01 * dot12 - dot11 * dot02) * dot03
+    w += (dot01 * dot12 - dot00 * dot12) * dot13
+    w += (dot00 * dot11 - dot01 * dot01) * dot23
+    w /= det
+
+    return u >= 0 and v >= 0 and w >= 0 and u + v + w <= 1
