+++ symfem/elements/guzman_neilan.py
@@ -34,13 +34,13 @@ class GuzmanNeilan(CiarletElement):
 
         dofs += make_integral_moment_dofs(
             reference,
-            facets=(NormalIntegralMoment, DiscontinuousLagrange, 0, "contravariant",
-                    {"variant": "equispaced"}),
+            facets=(NormalIntegralMoment, DiscontinuousLagrange, 0, "contravariant"),
         )
 
-        for dir in [(1, 0), (0, 1)]:
-            dofs.append(DotPointEvaluation((sympy.Rational(1, 3),
-                                            sympy.Rational(1, 3)), dir, entity=(2, 0)))
+        mid = tuple(sympy.Rational(sum(i), len(i)) for i in zip(*reference.vertices))
+        for i in range(reference.tdim):
+            dir = tuple(1 if i == j else 0 for j in range(reference.tdim))
+            dofs.append(DotPointEvaluation(mid, dir, entity=(reference.tdim, 0)))
 
         super().__init__(reference, order, poly, dofs, reference.tdim, reference.tdim)
 
@@ -57,27 +57,34 @@ class GuzmanNeilan(CiarletElement):
 
         refs = [create_reference("triangle", vs) for vs in sub_tris]
 
+        basis = []
+
         lagrange_spaces = [VectorDiscontinuousLagrange(ref, 1) for ref in refs]
-
-        piece_list = []
-        for i in range(2):
-            piece_list.append([lagrange_spaces[0].get_basis_function(i), (0, 0),
-                               lagrange_spaces[2].get_basis_function(2 + i)])
-            piece_list.append([lagrange_spaces[0].get_basis_function(2 + i),
-                               lagrange_spaces[1].get_basis_function(i), (0, 0)])
-            piece_list.append([(0, 0), lagrange_spaces[1].get_basis_function(2 + i),
-                               lagrange_spaces[2].get_basis_function(i)])
-            piece_list.append([lagrange_spaces[0].get_basis_function(4 + i),
-                               lagrange_spaces[1].get_basis_function(4 + i),
-                               lagrange_spaces[2].get_basis_function(4 + i)])
-
-        basis = [
-            PiecewiseFunction(list(zip(sub_tris, p)))
-            for p in piece_list
-        ]
+        for i in [
+            (0, None, 2), (1, None, 3),
+            (2, 0, None), (3, 1, None),
+            (None, 2, 0), (None, 3, 1),
+            (4, 4, 4), (5, 5, 5)
+        ]:
+            basis.append(
+                PiecewiseFunction(list(zip(sub_tris, [
+                    (0, 0) if j is None else s.get_basis_function(j)
+                    for s, j in zip(lagrange_spaces, i)
+                ])))
+            )
 
         fs = create_element("triangle", "Bernardi-Raugel", 1).get_basis_functions()[-3:]
-        sub_p2s = [VectorDiscontinuousLagrange(ref, 2, "equispaced") for ref in refs]
+        sub_p2s = [VectorDiscontinuousLagrange(ref, 2) for ref in refs]
+        sub_basis = []
+        for i in [
+            (10, 10, 10), (11, 11, 11),
+            (8, 6, None), (9, 7, None),
+            (6, None, 8), (7, None, 9),
+            (None, 8, 6), (None, 9, 7)
+        ]:
+            sub_fs = [(0, 0) if j is None else s.get_basis_functions()[j]
+                      for s, j in zip(sub_p2s, i)]
+            sub_basis.append(PiecewiseFunction([(tri, f) for f, tri in zip(sub_fs, sub_tris)]))
 
         for f in fs:
             fun = (div(f) - reference.integral(
[…]
