### 2021-07-25 #3514 @iammix
![](https://github.trello.services/images/mini-trello-icon.png) [[REVIEW] SymFEM Python Joss Journal](https://trello.com/c/U7R5DLOh/3-review-symfem-python-joss-journal)
