### 2021-07-28 261fe3e
add CITATION.cff
