+++ ++version.py
@@ -66,4 +66,17 @@ with open(".github/workflows/test-packages.yml") as f:
 with open(".github/workflows/test-packages.yml", "w") as f:
     f.write(new_test)
 
+# CITATION.cff
+new_citation = ""
+with open("CITATION.cff") as f:
+    for line in f:
+        if line.startswith("version: "):
+            new_citation += f"version: {new_version_str}\n"
+        elif line.startswith("date-released: "):
+            new_citation += f"date-released: {now.strftime('%Y-%m-%d')}\n"
+        else:
+            new_citation += line
+with open("CITATION.cff", "w") as f:
+    f.write(new_citation)
+
 print(f"Updated version to {new_version_str}")
