+++ symfem/elements/nedelec.py
@@ -29,6 +29,10 @@ class NedelecFirstKind(CiarletElement):
         )
 
         super().__init__(reference, order, poly, dofs, reference.tdim, reference.tdim)
+        self.variant = variant
+
+    def init_kwargs(self):
+        return {"variant": self.variant}
 
     names = ["Nedelec", "Nedelec1", "N1curl"]
     references = ["triangle", "tetrahedron"]
@@ -50,6 +54,10 @@ class NedelecSecondKind(CiarletElement):
         )
 
         super().__init__(reference, order, poly, dofs, reference.tdim, reference.tdim)
+        self.variant = variant
+
+    def init_kwargs(self):
+        return {"variant": self.variant}
 
     names = ["Nedelec2", "N2curl"]
     references = ["triangle", "tetrahedron"]
