+++ symfem/elements/mtw.py
@@ -56,6 +56,7 @@ class MardalTaiWinther(CiarletElement):
         self.variant = variant
 
     def init_kwargs(self):
+        """Return the kwargs used to create this element."""
         return {"variant": self.variant}
 
     names = ["Mardal-Tai-Winther", "MTW"]
