+++ symfem/elements/nedelec_prism.py
@@ -53,6 +53,7 @@ class Nedelec(CiarletElement):
         self.variant = variant
 
     def init_kwargs(self):
+        """Return the kwargs used to create this element."""
         return {"variant": self.variant}
 
     names = ["Nedelec", "Ncurl"]
