+++ symfem/elements/bddm.py
@@ -47,6 +47,7 @@ class BDDF(CiarletElement):
         super().__init__(reference, order, poly, dofs, reference.tdim, reference.tdim)
 
     def init_kwargs(self):
+        """Return the kwargs used to create this element."""
         return {"variant": self.variant}
 
     names = ["Brezzi-Douglas-Duran-Fortin", "BDDF"]
