+++ symfem/elements/aw.py
@@ -67,6 +67,7 @@ class ArnoldWinther(CiarletElement):
                          (reference.tdim, reference.tdim))
 
     def init_kwargs(self):
+        """Return the kwargs used to create this element."""
         return {"variant": self.variant}
 
     names = ["Arnold-Winther", "AW"]
