+++ symfem/elements/bdfm.py
@@ -60,6 +60,7 @@ class BDFM(CiarletElement):
         super().__init__(reference, order, poly, dofs, reference.tdim, reference.tdim)
 
     def init_kwargs(self):
+        """Return the kwargs used to create this element."""
         return {"variant": self.variant}
 
     names = ["Brezzi-Douglas-Fortin-Marini", "BDFM"]
