### 2021-07-31 f859431
convert Basis Functions to sympy in matrix
