+++ symfem/elements/guzman_neilan.py
@@ -17,6 +17,9 @@ class GuzmanNeilan(CiarletElement):
     """Guzman-Neilan Hdiv finite element."""
 
     def __init__(self, reference, order):
+        if reference.name == "tetrahedron":
+            raise NotImplementedError()
+
         if reference.name == "triangle":
             poly = self._make_polyset_triangle(reference, order)
         else:
@@ -109,190 +112,65 @@ class GuzmanNeilan(CiarletElement):
 
         coeffs = [
             [
-                -sympy.Rational(5, 2),
-                -sympy.Rational(5, 8),
-                sympy.Rational(25, 8),
-                sympy.Rational(5, 18),
-                -sympy.Rational(35, 72),
-                sympy.Rational(5, 24),
-                sympy.Rational(1, 3),
-                -sympy.Rational(7, 9),
-                sympy.Rational(13, 9),
-                sympy.Rational(-21061, 793152),
-                sympy.Rational(-1833299, 793152),
-                sympy.Rational(275, 72),
-                sympy.Rational(-109189, 396576),
-                sympy.Rational(-1756187, 396576),
-                sympy.Rational(22, 3),
-                sympy.Rational(-1153979, 793152),
-                sympy.Rational(25, 72),
-                sympy.Rational(291059, 793152),
-                sympy.Rational(-1242107, 396576),
-                -sympy.Rational(4, 9),
-                sympy.Rational(698651, 396576),
-                -sympy.Rational(25, 18),
-                sympy.Rational(1429379, 793152),
-                sympy.Rational(-915299, 793152),
-                -sympy.Rational(13, 3),
-                sympy.Rational(1506491, 396576),
-                sympy.Rational(-507707, 396576),
-                sympy.Rational(-1145819, 176256),
-                sympy.Rational(8586531, 1586304),
-                sympy.Rational(-5, 216),
-                sympy.Rational(-933659, 1586304),
-                -sympy.Rational(655, 216),
-                sympy.Rational(456299, 1586304),
-                -sympy.Rational(185, 999),
-                sympy.Rational(1043819, 1586304),
-                sympy.Rational(-750059, 1586304),
-                sympy.Rational(1162139, 176256),
-                sympy.Rational(25, 216),
-                sympy.Rational(-1182539, 176256),
-                sympy.Rational(199259, 1586304),
-                sympy.Rational(-2218859, 1586304),
-                sympy.Rational(275, 216),
-                -sympy.Rational(10, 27),
-                sympy.Rational(-546059, 176256),
-                sympy.Rational(611339, 176256)
-            ], [
-                sympy.Rational(5, 1),
-                sympy.Rational(-5, 8),
-                sympy.Rational(-35, 8),
-                sympy.Rational(-10, 27),
-                sympy.Rational(115, 216),
-                sympy.Rational(-35, 216),
-                sympy.Rational(1, 27),
-                sympy.Rational(16, 27),
-                sympy.Rational(-44, 27),
-                sympy.Rational(-766499, 396576),
-                sympy.Rational(2042519, 396576),
-                sympy.Rational(-1175, 216),
-                sympy.Rational(-612275, 198288),
-                sympy.Rational(1948883, 198288),
-                sympy.Rational(-293, 27),
-                sympy.Rational(1133699, 396576),
-                sympy.Rational(-425, 216),
-                sympy.Rational(-353399, 396576),
-                sympy.Rational(1287923, 198288),
[…]
