+++ symfem/elements/guzman_neilan.py
@@ -17,9 +17,6 @@ class GuzmanNeilan(CiarletElement):
     """Guzman-Neilan Hdiv finite element."""
 
     def __init__(self, reference, order):
-        if reference.name == "tetrahedron":
-            raise NotImplementedError()
-
         if reference.name == "triangle":
             poly = self._make_polyset_triangle(reference, order)
         else:
