+++ symfem/elements/guzman_neilan.py
@@ -74,17 +74,7 @@ class GuzmanNeilan(CiarletElement):
         """Make the polyset for a triangle."""
         assert order == 1
 
-        coeffs = [
-            [sympy.Rational(1, 6), sympy.Rational(1, 6), sympy.Rational(5, 24),
-             sympy.Rational(-1, 24), sympy.Rational(5, 24), sympy.Rational(5, 24),
-             sympy.Rational(-1, 24), sympy.Rational(5, 24)],
-            [sympy.Rational(1, 3), sympy.Rational(-2, 3), sympy.Rational(5, 12),
-             sympy.Rational(-1, 3), sympy.Rational(-1, 12), sympy.Rational(-1, 3),
-             sympy.Rational(5, 12), sympy.Rational(-5, 6)],
-            [sympy.Rational(2, 3), sympy.Rational(-1, 3), sympy.Rational(5, 6),
-             sympy.Rational(-5, 12), sympy.Rational(1, 3), sympy.Rational(1, 12),
-             sympy.Rational(1, 3), sympy.Rational(-5, 12)]
-        ]
+        from ._guzman_neilan_triangle import coeffs
 
         mid = tuple(sympy.Rational(sum(i), len(i)) for i in zip(*reference.vertices))
 
@@ -106,69 +96,8 @@ class GuzmanNeilan(CiarletElement):
     def _make_polyset_tetrahedron(self, reference, order):
         """Make the polyset for a tetrahedron."""
         assert order in [1, 2]
+        from ._guzman_neilan_tetrahedron import coeffs
 
-        coeffs = [
-            [
-                0, 0, 0, sympy.Rational(1, 3), sympy.Rational(1, 3), sympy.Rational(1, 3),
-                sympy.Rational(-121, 1620), sympy.Rational(-1379, 1620), sympy.Rational(25, 27),
-                sympy.Rational(149, 810), sympy.Rational(-1109, 810), sympy.Rational(23, 27),
-                sympy.Rational(-1379, 1620), sympy.Rational(25, 27), sympy.Rational(-121, 1620),
-                sympy.Rational(-1109, 810), sympy.Rational(23, 27), sympy.Rational(149, 810),
-                sympy.Rational(25, 27), sympy.Rational(-121, 1620), sympy.Rational(-1379, 1620),
-                sympy.Rational(23, 27), sympy.Rational(149, 810), sympy.Rational(-1109, 810),
-                sympy.Rational(-679, 360), sympy.Rational(193, 120), sympy.Rational(-5, 54),
-                sympy.Rational(-1379, 3240), sympy.Rational(-35, 54), sympy.Rational(-121, 3240),
-                sympy.Rational(-35, 54), sympy.Rational(-121, 3240), sympy.Rational(-1379, 3240),
-                sympy.Rational(193, 120), sympy.Rational(-5, 54), sympy.Rational(-679, 360),
-                sympy.Rational(-121, 3240), sympy.Rational(-1379, 3240), sympy.Rational(-35, 54),
-                sympy.Rational(-5, 54), sympy.Rational(-679, 360), sympy.Rational(193, 120),
-            ],
-            [
-                0, 0, 0, sympy.Rational(-1, 3), sympy.Rational(-1, 3), sympy.Rational(-1, 3),
-                sympy.Rational(-1087, 810), sympy.Rational(1987, 810), sympy.Rational(-25, 18),
-                sympy.Rational(-1222, 405), sympy.Rational(1852, 405), sympy.Rational(-16, 9),
-                sympy.Rational(1087, 810), sympy.Rational(-25, 18), sympy.Rational(-187, 810),
-                sympy.Rational(952, 405), sympy.Rational(-16, 9), sympy.Rational(-322, 405),
[…]
