+++ test/test_guzman_neilan.py
@@ -3,6 +3,8 @@ import symfem
 import sympy
 from itertools import combinations
 from symfem.calculus import div
+from symfem.symbolic import subs, x
+from symfem.elements.guzman_neilan import make_piecewise_lagrange
 
 
 @pytest.mark.parametrize("order", [1])
@@ -11,6 +13,7 @@ def test_guzman_neilan_triangle(order):
 
     for p in e._basis[-3:]:
         for piece in p.pieces:
+            print(div(piece[1]).expand())
             float(div(piece[1]).expand())
 
 
@@ -18,9 +21,13 @@ def test_guzman_neilan_triangle(order):
 def test_guzman_neilan_tetrahedron(order):
     e = symfem.create_element("tetrahedron", "Guzman-Neilan", order)
 
+    mid = tuple(sympy.Rational(sum(i), len(i)) for i in zip(*e.reference.vertices))
     for p in e._basis[-4:]:
         for piece in p.pieces:
+            print(div(piece[1]).expand())
             float(div(piece[1]).expand())
+
+        assert subs(p, x, mid) == (0, 0, 0)
 
 
 @pytest.mark.parametrize("order", [1])
@@ -53,9 +60,9 @@ def test_basis_continuity_triangle(order):
 def test_basis_continuity_tetrahedron(order):
     N = 5
     e = symfem.create_element("tetrahedron", "Guzman-Neilan", order)
-    sixth = sympy.Rational(1, 6)
+    quarter = sympy.Rational(1, 4)
     one = sympy.Integer(1)
-    for pt in [(0, 0, 0), (1, 0, 0), (0, 1, 0), (0, 0, 1), (sixth, sixth, sixth)]:
+    for pt in [(0, 0, 0), (1, 0, 0), (0, 1, 0), (0, 0, 1), (quarter, quarter, quarter)]:
         for f in e.get_polynomial_basis():
             value = None
             for p in f.pieces:
@@ -63,14 +70,79 @@ def test_basis_continuity_tetrahedron(order):
                     if value is None:
                         value = symfem.symbolic.subs(p[1], symfem.symbolic.x, pt)
                     assert value == symfem.symbolic.subs(p[1], symfem.symbolic.x, pt)
-    for pts in combinations([(0, 0, 0), (1, 0, 0), (0, 1, 0), (0, 0, 1), (sixth, sixth, sixth)], 2):
+    for pts in combinations([(0, 0, 0), (1, 0, 0), (0, 1, 0), (0, 0, 1),
+                             (quarter, quarter, quarter)], 2):
         for i in range(N + 1):
             pt = tuple(a + (b - a) * i * one / N for a, b in zip(*pts))
             for f in e.get_polynomial_basis():
-                print(f.pieces)
                 value = None
                 for p in f.pieces:
                     if pts[0] in p[0] and pts[1] in p[0]:
                         if value is None:
                             value = symfem.symbolic.subs(p[1], symfem.symbolic.x, pt)
+                        if value != symfem.symbolic.subs(p[1], symfem.symbolic.x, pt):
+                            from IPython import embed; embed()
                         assert value == symfem.symbolic.subs(p[1], symfem.symbolic.x, pt)
+    for pts in combinations([(0, 0, 0), (1, 0, 0), (0, 1, 0), (0, 0, 1),
+                             (quarter, quarter, quarter)], 3):
+        for i in range(N + 1):
+            for j in range(N + 1 - i):
+                pt = tuple(a + (b - a) * i * one / N + (c - a) * j * one / N
+                           for a, b, c in zip(*pts))
+                for f in e.get_polynomial_basis():
+                    value = None
+                    for p in f.pieces:
[…]
