+++ symfem/elements/_guzman_neilan_triangle.py
@@ -1,0 +1,16 @@
+import sympy
+
+coeffs = [
+    [
+        sympy.Rational(1, 6), sympy.Rational(1, 6), sympy.Rational(5, 24), sympy.Rational(-1, 24),
+        sympy.Rational(5, 24), sympy.Rational(5, 24), sympy.Rational(-1, 24), sympy.Rational(5, 24),
+    ],
+    [
+        sympy.Rational(1, 3), sympy.Rational(-2, 3), sympy.Rational(5, 12), sympy.Rational(-1, 3),
+        sympy.Rational(-1, 12), sympy.Rational(-1, 3), sympy.Rational(5, 12), sympy.Rational(-5, 6),
+    ],
+    [
+        sympy.Rational(2, 3), sympy.Rational(-1, 3), sympy.Rational(5, 6), sympy.Rational(-5, 12),
+        sympy.Rational(1, 3), sympy.Rational(1, 12), sympy.Rational(1, 3), sympy.Rational(-5, 12),
+    ],
+]

[…]
