### 2021-08-01 c0fdc4a
fix midpoint bug
