+++ test/test_elements.py
@@ -204,7 +204,10 @@ def test_element_functionals_and_continuity(
     for dim, entities in entity_pairs:
         for fi, gi in zip(*[space.entity_dofs(dim, i) for i in entities]):
             basis = space.get_basis_functions()
-            basis2 = space.map_to_cell(vertices)
+            try:
+                basis2 = space.map_to_cell(vertices)
+            except NotImplementedError:
+                pytest.xfail("Mapping not implemented for this element.")
 
             f = basis[fi]
             g = basis2[gi]
