### 2021-08-01 8b250ff
Add Guzman-Neilan element (#112)
* Add Guzman-Neilan on a triangle

* working on Guzman-Neilan element on tetrahedron

* pydocstyle

* working on Guzman-Neilan element

* update docs

* docs

* init kwargs

* tidy up

* fix various errors

* fix transition element

* remove prints

* convert Basis Functions to sympy in matrix

* try a better basis

* start writing more test for Guzman-Neilan

* working on Guzman-Neilan

* flake

* fix midpoint bug

* add back xfail

* 1/4 not 1/3
