+++ symfem/elements/_guzman_neilan_tetrahedron.py
@@ -1,0 +1,69 @@
+import sympy
+
+coeffs = [
+    [
+        sympy.Rational(5, 8), sympy.Rational(5, 8), sympy.Rational(5, 8), 0, 0,
+        sympy.Rational(125, 72), sympy.Rational(17, 36), sympy.Rational(17, 36),
+        sympy.Rational(71, 18), sympy.Rational(-6343, 81000), sympy.Rational(-16157, 81000),
+        sympy.Rational(325, 216), sympy.Rational(6391, 20250), sympy.Rational(742, 10125),
+        sympy.Rational(58, 27), sympy.Rational(-7879, 10125), sympy.Rational(325, 216),
+        sympy.Rational(-100093, 81000), sympy.Rational(-43907, 40500), sympy.Rational(58, 27),
+        sympy.Rational(-20242, 10125), sympy.Rational(325, 216), sympy.Rational(-26609, 40500),
+        sympy.Rational(-109907, 81000), sympy.Rational(58, 27), sympy.Rational(-34093, 40500),
+        sympy.Rational(-45391, 20250), sympy.Rational(-13471, 54000), sympy.Rational(15971, 54000),
+        sympy.Rational(-25, 27), sympy.Rational(-7879, 20250), sympy.Rational(-545, 144),
+        sympy.Rational(10133, 40500), sympy.Rational(-545, 144), sympy.Rational(-26609, 81000),
+        sympy.Rational(15359, 81000), sympy.Rational(39149, 13500), sympy.Rational(-25, 432),
+        sympy.Rational(-9631, 3375), sympy.Rational(-6343, 162000), sympy.Rational(-16157, 162000),
+        sympy.Rational(145, 36), sympy.Rational(-25, 432), sympy.Rational(63577, 27000),
+        sympy.Rational(-62327, 27000),
+    ],
+    [
+        sympy.Rational(15, 8), 0, 0, 0, sympy.Rational(125, 72), 0, sympy.Rational(1, 12),
+        sympy.Rational(113, 36), sympy.Rational(-1, 3), sympy.Rational(-10921, 8100),
+        sympy.Rational(27467, 16200), sympy.Rational(-25, 18), sympy.Rational(-21167, 8100),
+        sympy.Rational(24767, 8100), sympy.Rational(-16, 9), sympy.Rational(25217, 16200),
+        sympy.Rational(-25, 18), sympy.Rational(8533, 16200), sympy.Rational(6473, 2025),
+        sympy.Rational(-16, 9), sympy.Rational(5833, 8100), sympy.Rational(125, 72),
+        sympy.Rational(-8273, 4050), sympy.Rational(13967, 16200), sympy.Rational(44, 9),
+        sympy.Rational(-8948, 2025), sympy.Rational(11267, 8100), sympy.Rational(9671, 1800),
+        sympy.Rational(-2449, 450), sympy.Rational(25, 36), sympy.Rational(25217, 32400),
+        sympy.Rational(235, 48), sympy.Rational(8533, 32400), sympy.Rational(-235, 144),
+        sympy.Rational(-4967, 32400), sympy.Rational(13967, 32400), sympy.Rational(-9239, 1200),
+        sympy.Rational(-25, 144), sympy.Rational(27467, 3600), sympy.Rational(-10921, 16200),
+        sympy.Rational(6949, 4050), sympy.Rational(-35, 12), sympy.Rational(-145, 144),
+        sympy.Rational(12967, 3600), sympy.Rational(-663, 400),
+    ],
+    [
+        0, sympy.Rational(-15, 8), 0, 0, sympy.Rational(-61387, 81000),
[…]
