### 2021-08-01 39032e0
allow for floats in tetrahedron check
