### 2021-08-02 #3556 @Nikoleta-v3
👋🏼  @mscroggs & @iammix & @sigvaldm this is the review thread for the paper. All of our communications will happen here from now on.

Both reviewers have checklists at the top of this thread with the JOSS requirements 🔝 As you go over the submission, please check any items that you feel have been satisfied. There are also links to the JOSS reviewer guidelines.

Please feel free to ping me (@Nikoleta-v3 ) if you have any questions/concerns 👍🏻 😄
