### 2021-08-02 a3d0ecc
add conforming Arnold-Winther to docs
