+++ symfem/create.py
@@ -146,7 +146,7 @@ def create_element(cell_type, element_type, order, **kwargs):
         Brezzi-Douglas-Fortin-Marini, BDFM,
         Brezzi-Douglas-Duran-Fortin, BDDF,
         Hellan-Herrmann-Johnson, HHJ,
-        Arnold-Winther, AW,
+        Arnold-Winther, AW, conforming Arnold-Winther,
         Bell,
         Kong-Mulder-Veldhuizen, KMV,
         Bernstein, Bernstein-Bezier,
