### 2021-08-06 3747c22
Move testing to FiniteElement class
