+++ test/test_elements.py
@@ -1,13 +1,8 @@
 import pytest
-import sympy
 import symfem
 from symfem import create_element
-from symfem.finite_element import CiarletElement, DirectElement
-from symfem.symbolic import subs, x, PiecewiseFunction
-from symfem.vectors import vsub
-from symfem.symbolic import to_sympy
-from symfem.calculus import diff
-from .utils import test_elements, all_symequal
+from symfem.symbolic import subs, x, all_symequal
+from .utils import test_elements
 
 
 def test_all_tested():
@@ -45,11 +40,11 @@ def test_too_low_order(ref, element, order):
     [[reference, element, order, kwargs]
      for reference, i in test_elements.items() for element, j in i.items()
      for kwargs, k in j for order in k])
-def test_independence(
+def test_element(
     elements_to_test, cells_to_test, cell_type, element_type, order, kwargs,
     speed
 ):
-    """Test that DirectElements have independent basis functions."""
+    """Run tests for each element."""
     if elements_to_test != "ALL" and element_type not in elements_to_test:
         pytest.skip()
     if cells_to_test != "ALL" and cell_type not in cells_to_test:
@@ -61,223 +56,7 @@ def test_independence(
             pytest.skip()
 
     space = create_element(cell_type, element_type, order, **kwargs)
-
-    # Only run this test for DirectElements
-    if not isinstance(space, DirectElement):
-        pytest.skip()
-
-    basis = space.get_basis_functions()
-    all_terms = set()
-
-    try:
-        basis[0].as_coefficients_dict()
-        scalar = True
-    except AttributeError:
-        scalar = False
-
-    if scalar:
-        for f in basis:
-            for term in f.as_coefficients_dict():
-                all_terms.add(term)
-        mat = [[0 for i in all_terms] for j in basis]
-        for i, t in enumerate(all_terms):
-            for j, f in enumerate(basis):
-                fd = f.as_coefficients_dict()
-                if t in fd:
-                    mat[j][i] = fd[t]
-    else:
-        for f in basis:
-            for fi, fpart in enumerate(f):
-                for term in fpart.as_coefficients_dict():
-                    all_terms.add((fi, term))
-        mat = [[0 for i in all_terms] for j in basis]
-        for i, (fi, t) in enumerate(all_terms):
-            for j, f in enumerate(basis):
-                fd = f[fi].as_coefficients_dict()
-                if t in fd:
-                    mat[j][i] = fd[t]
-    mat = sympy.Matrix(mat)
-
-    assert mat.rank() == mat.rows
-
-
-@pytest.mark.parametrize(
-    ("cell_type", "element_type", "order", "kwargs"),
-    [[reference, element, order, kwargs]
-     for reference, i in test_elements.items() for element, j in i.items()
-     for kwargs, k in j for order in k])
-def test_functional_entities(
-    elements_to_test, cells_to_test, cell_type, element_type, order, kwargs,
-    speed
-):
-    if elements_to_test != "ALL" and element_type not in elements_to_test:
-        pytest.skip()
-    if cells_to_test != "ALL" and cell_type not in cells_to_test:
-        pytest.skip()
-
-    # Test functionals
-    space = create_element(cell_type, element_type, order, **kwargs)
-    if not isinstance(space, CiarletElement):
-        pytest.skip()
-    for dof in space.dofs:
-        dim, entity = dof.entity
-        assert entity < space.reference.sub_entity_count(dim)
-        if hasattr(dof, "reference"):
-            assert dim == dof.reference.tdim
-
-
[…]
