+++ test/utils.py
@@ -1,7 +1,5 @@
 """Utility functions for testing."""
 
-import sympy
-
 test_elements = {
     "interval": {
         "P": [[{"variant": "equispaced"}, range(6)], [{"variant": "lobatto"}, range(4)]],
@@ -164,13 +162,3 @@ test_elements = {
         "Lagrange": [[{"variant": "equispaced"}, range(4)]],
     }
 }
-
-
-def all_symequal(a, b):
-    """Check if two symbolic numbers or vectors are equal."""
-    if isinstance(a, (list, tuple)):
-        for i, j in zip(a, b):
-            if not all_symequal(i, j):
-                return False
-        return True
-    return sympy.expand(sympy.simplify(a)) == sympy.expand(sympy.simplify(b))
