+++ symfem/finite_element.py
@@ -3,7 +3,9 @@
 import sympy
 import warnings
 import numpy
-from .symbolic import x, subs, sym_sum, PiecewiseFunction, to_sympy, to_float
+from .symbolic import x, subs, sym_sum, PiecewiseFunction, to_sympy, to_float, all_symequal
+from .calculus import diff
+from .vectors import vsub
 from .basis_function import ElementBasisFunction
 from .legendre import evaluate_legendre_basis, get_legendre_basis
 
@@ -79,6 +81,132 @@ class FiniteElement:
     def map_to_cell(self, vertices, basis=None):
         """Map the basis onto a cell using the appropriate mapping for the element."""
         raise NotImplementedError()
+
+    def test(self):
+        """Run tests for this element."""
+        if self.order <= 4:
+            self.test_continuity()
+
+    def test_continuity(self):
+        """Test that this element has the correct continuity."""
+        continuity = self.continuity
+        if "{order}" in continuity:
+            continuity = continuity.replace("{order}", f"{self.order}")
+
+        # Test continuity
+        if self.reference.name == "interval":
+            vertices = ((-1, ), (0, ))
+            entity_pairs = [[0, (0, 1)]]
+        elif self.reference.name == "triangle":
+            vertices = ((-1, 0), (0, 0), (0, 1))
+            entity_pairs = [[0, (0, 1)], [0, (2, 2)], [1, (1, 0)]]
+        elif self.reference.name == "tetrahedron":
+            vertices = ((-1, 0, 0), (0, 0, 0), (0, 1, 0), (0, 0, 1))
+            entity_pairs = [[0, (0, 1)], [0, (2, 2)], [0, (3, 3)],
+                            [1, (0, 0)], [1, (3, 1)], [1, (4, 2)],
+                            [2, (1, 0)]]
+        elif self.reference.name == "quadrilateral":
+            vertices = ((0, 0), (0, 1), (-1, 0), (-1, 1))
+            entity_pairs = [[0, (0, 0)], [0, (2, 1)], [1, (1, 0)]]
+        elif self.reference.name == "hexahedron":
+            vertices = ((0, 0, 0), (0, 0, 1), (0, 1, 0), (0, 1, 1),
+                        (-1, 0, 0), (-1, 0, 1), (-1, 1, 0), (-1, 1, 1))
+            entity_pairs = [[0, (0, 0)], [0, (2, 2)], [0, (4, 1)], [0, (6, 3)],
+                            [1, (1, 1)], [1, (2, 0)], [1, (6, 5)], [1, (9, 3)],
+                            [2, (0, 2)]]
+        elif self.reference.name == "prism":
+            vertices = ((-1, 0, 0), (0, 0, 0), (0, 1, 0),
+                        (-1, 0, 1), (0, 0, 1), (0, 1, 1))
+            entity_pairs = [[0, (0, 1)], [0, (2, 2)], [0, (3, 4)], [0, (5, 5)],
+                            [1, (1, 3)], [1, (2, 4)], [1, (6, 6)], [1, (7, 8)],
+                            [2, (2, 3)]]
+        elif self.reference.name == "pyramid":
+            vertices = ((-1, 0, 0), (0, 0, 0), (-1, 1, 0),
+                        (0, 1, 0), (0, 0, 1))
+            entity_pairs = [[0, (0, 1)], [0, (2, 3)], [0, (4, 4)],
+                            [1, (1, 3)], [1, (2, 4)], [1, (6, 7)],
+                            [2, (2, 3)]]
+
+        if continuity == "L2":
+            return
+
+        for dim, entities in entity_pairs:
+            for fi, gi in zip(*[self.entity_dofs(dim, i) for i in entities]):
+                basis = self.get_basis_functions()
+                try:
+                    basis2 = self.map_to_cell(vertices)
[…]
