+++ .github/ISSUE_TEMPLATE/support.md
@@ -1,0 +1,8 @@
+---
+name: Support
+about: Ask a question about using Symfem
+title: ''
+labels: support
+assignees: ''
+
+---
