### 2021-08-09 #3556 @mscroggs
> Not yet done with the review, but my first impression is good. There are clear instructions on how to 1) contribute and 2) report issues, but not 3) seek support. One could perhaps append a statement in the "Using Symfem" section that in addition to the documentation, further support may be reached through e-mail, or the issue tracker, or whatever your preference is.

Good idea, I've added a link to opening a support issue in the Using Symfem section
