### 2021-08-09 27397f9
CITATION.md is in the root dir
