### 2021-08-09 c175f5c
Update Basix interface (#115)
* Update Basix interface

* CITATION.md is in the root dir
