+++ test/test_against_basix.py
@@ -3,41 +3,45 @@ import numpy as np
 import pytest
 
 elements = {
-    "interval": [("P", "Lagrange", range(1, 4)), ("dP", "Discontinuous Lagrange", range(1, 4)),
-                 ("serendipity", "Serendipity", range(1, 5)),
-                 ("bubble", "Bubble", range(2, 5)),
-                 ("dP", "DPC", range(0, 5))],
-    "triangle": [("P", "Lagrange", range(1, 4)), ("dP", "Discontinuous Lagrange", range(1, 4)),
-                 ("bubble", "Bubble", range(3, 5)),
-                 ("N1curl", "Nedelec 1st kind H(curl)", range(1, 4)),
-                 ("N2curl", "Nedelec 2nd kind H(curl)", range(1, 4)),
-                 ("N1div", "Raviart-Thomas", range(1, 4)),
-                 ("N2div", "Brezzi-Douglas-Marini", range(1, 4)),
-                 ("Regge", "Regge", range(0, 4)),
-                 ("Crouzeix-Raviart", "Crouzeix-Raviart", [1])],
-    "tetrahedron": [("P", "Lagrange", range(1, 4)), ("dP", "Discontinuous Lagrange", range(1, 4)),
-                    ("bubble", "Bubble", range(4, 6)),
-                    ("N1curl", "Nedelec 1st kind H(curl)", range(1, 4)),
-                    ("N2curl", "Nedelec 2nd kind H(curl)", range(1, 3)),
-                    ("N1div", "Raviart-Thomas", range(1, 3)),
-                    ("N2div", "Brezzi-Douglas-Marini", range(1, 3)),
-                    ("Regge", "Regge", range(0, 3)),
-                    ("Crouzeix-Raviart", "Crouzeix-Raviart", [1])],
-    "quadrilateral": [("Q", "Lagrange", range(1, 4)),
-                      ("dQ", "Discontinuous Lagrange", range(1, 4)),
-                      ("dP", "DPC", range(0, 4)),
-                      ("serendipity", "Serendipity", range(1, 5)),
-                      ("Qdiv", "Raviart-Thomas", range(1, 4)),
-                      ("Qcurl", "Nedelec 1st kind H(curl)", range(1, 4)),
-                      ("Sdiv", "Brezzi-Douglas-Marini", range(1, 4)),
-                      ("Scurl", "Nedelec 2nd kind H(curl)", range(1, 4))],
-    "hexahedron": [("Q", "Lagrange", range(1, 3)), ("dQ", "Discontinuous Lagrange", range(1, 3)),
-                   ("serendipity", "Serendipity", range(1, 5)),
-                   ("Qdiv", "Raviart-Thomas", range(1, 3)),
-                   ("Qcurl", "Nedelec 1st kind H(curl)", range(1, 3)),
-                   ("Sdiv", "Brezzi-Douglas-Marini", range(1, 3)),
-                   ("Scurl", "Nedelec 2nd kind H(curl)", range(1, 3))],
-    "prism": [("Lagrange", "Lagrange", range(1, 4))]
+    "interval": [("P", "Lagrange", range(1, 4), ["equispaced"]),
+                 ("dP", "Discontinuous Lagrange", range(1, 4), []),
+                 ("serendipity", "Serendipity", range(1, 5), []),
+                 ("bubble", "Bubble", range(2, 5), []),
+                 ("dP", "DPC", range(0, 5), [])],
+    "triangle": [("P", "Lagrange", range(1, 4), ["equispaced"]),
+                 ("dP", "Discontinuous Lagrange", range(1, 4), []),
+                 ("bubble", "Bubble", range(3, 5), []),
[…]
