### 2021-08-10 #3556 @sigvaldm
General comment: Although there is no lack of software which implements various finite elements, Symfem implements a rather vast set of element types, and does so using symbolic representations (using SymPy, at increased computational cost). This allows further symbolic manipulation (differentiation and so on) by the user. The software is well designed, using the general Ciarlet definition (`CiarletElement`) of the finite element as the starting point for most (all?) element types, and allowing new types of elements to be implemented through inheritance (one file per element). This makes the software extensible and maintainable. The software is also well tested through continuous integration and parametrized tests. This enforces the maintainability and generality, and it does not appear that the tests are too tightly coupled to implementation. Some of the tests were skipped on my system. The API seems to be well complemented by docstrings, although the introductory documentation is perhaps somewhat terse. Although the software is still young, and has only a sole developer (except for a bot), it is my opinion that the work represents substantial scholarly effort, and can make it easier to experimenting with various elements during development also for other researchers. I will make more specific comments and requests below.
