### 2021-08-10 #3556 @sigvaldm
While trying out the software I was a little confused by `create_reference()` at first, because it allowed specifying the coordinates of the vertices. What would distinguish a reference element from any other element then? The stiffness matrix example also used `create_reference()` to create all elements in the mesh, specifying the coordinates through the second argument. One might also expect `create_reference()` and `create_element()` to returns the same kind of object, except that the former is a reference element, but in fact `create_element()` returns an object with a higher level of abstraction, and it seems that `create_reference()` creates the lower level of abstraction regardless of the vertices' position. One way in which the returned object from `create_reference()` is truly a reference element and not just any element, is that it contains forward and inverse mapping functions to any element. I think I figured out how to use it, anyway, but perhaps I'm missing something. Would you mind enlightening me?
