### 2021-08-10 #3556 @sigvaldm
There's a curious oversight in the paper: There's no reference on the finite element method when the finite element method and the weak form is introduced.
