### 2021-08-10 #3556 @Nikoleta-v3
:wave @sigvaldm

> @Nikoleta-v3: There are two statement-of-need items in the checklist. How is the one under "Documentation" different from the one under "Software paper"? Is it just that it should be clear from the readme/readthedocs as well what the software does?

Yup! The `statement-of-need` should be clear in both the paper and in the package's documentation.
