### 2021-08-10 #3556 @sigvaldm
Sounds good, @mscroggs. All in all a very nice piece of software. Perhaps the main area of improvement in my opinion would be the documentation.

Each method and class is well documented with to-the-point docstrings, so this is good. The introductory part of the documentation is very brief, though, and doesn't even expose many of the categories of functionality present in the API. From the API I see there are vector operators, differential operators, mappings between reference elements and other elements, and whatnot. Since these aren't mentioned at all in examples (demos) or the introduction, one might in fact get the impression that the software does less than it actually does. I don't want to impose too many requirements on documentation style, since that is personal, but I would appreciate at least mentioning what exists somewhere either in the introduction, or in examples. The dot product is used in the code snippet on the matrix assembly demo, and one could for instance use the occasion to say that there's a whole range of such operations with a reference to the API documentation. Although e.g. `vdot()` is used in the snippet it's easy to miss if you're just glossing over it. On a similar account, I have also peeked at your encyclopedia over at defelement.com, and I think it's an excellent complement to Symfem. Perhaps you could add a link to that as well, in case the user needs some background information?

The closest thing to a statement-of-need I could find in the documentation is this sentence: "Symfem is a symbolic finite element definition library, that can be used to symbolically evaluate the basis functions of a finite element space.". It's short, and doesn't quite say why the software is needed. Perhaps append a few more sentences after this one, or add some sentences to the beginning of the readthedocs?

Finally, as for example of usage, I liked the stiffness matrix demo, but you don't really need symbolic evaluation to assemble a stiffness matrix. Can you think of another example to add which benefits more from the fact that expressions are symbolic? Perhaps you know of something from DefElement you can adapt?

I think that's all my comments to begin with. Looking forward to the responses :)
