### 2021-08-10 #3556 @mscroggs
> While trying out the software I was a little confused by `create_reference()` at first, because it allowed specifying the coordinates of the vertices. What would distinguish a reference element from any other element then? The stiffness matrix example also used `create_reference()` to create all elements in the mesh, specifying the coordinates through the second argument. One might also expect `create_reference()` and `create_element()` to returns the same kind of object, except that the former is a reference element, but in fact `create_element()` returns an object with a higher level of abstraction, and it seems that `create_reference()` creates the lower level of abstraction regardless of the vertices' position. One way in which the returned object from `create_reference()` is truly a reference element and not just any element, is that it contains forward and inverse mapping functions to any element. I think I figured out how to use it, anyway, but perhaps I'm missing something. Would you mind enlightening me?

`create_reference` creates a reference cell which contains functions to create the map to/from the default reference, geometry computation, etc. `create_element` creates a finite element on a reference cell.

Being able to define the vertices of a reference cell is a little confusing: originally I was planning to allow the user to create finite elements on a cell with any vertices, but later decided that this was probably an overcomplication. It's very useful in multiple place to be able to create these cells and use their geometry functions though.

I'm currently updating some of the docstrings: I'll add some more detail about this.
