### 2021-08-10 ce15190
Corrected typos in JOSS paper (#116)
