### 2021-08-10 #3556 @mscroggs
> There's a curious oversight in the paper: There's no reference on the finite element method when the finite element method and the weak form is introduced.

I've added a reference in https://github.com/mscroggs/symfem/commit/1398dfde9ea156d02856682de5edf33b93274c88
