### 2021-08-10 #3556 @mscroggs
> Concerning the automated tests, on a clean conda environment the tests against Basix were skipped because Basix were not installed. This is reasonable, and I see that they pass on the continuous integration system. There were also five skipped tests in `test_legendre.py` (see below). This is probably fine, but I'm curious why they are skipped?
>
> ```
>
> test_legendre.py::test_legendre[prism-4] PASSED                                                                        [ 76%]
> test_legendre.py::test_legendre[pyramid-1] SKIPPED (<Skipped instance>)                                                [ 78%]
> test_legendre.py::test_legendre[pyramid-2] SKIPPED (<Skipped instance>)                                                [ 80%]
> test_legendre.py::test_legendre[pyramid-3] SKIPPED (<Skipped instance>)                                                [ 82%]
> test_legendre.py::test_legendre[pyramid-4] SKIPPED (<Skipped instance>)                                                [ 84%]
> test_legendre.py::test_orthogonal[interval] PASSED                                                                     [ 86%]
> test_legendre.py::test_orthogonal[triangle] PASSED                                                                     [ 89%]
> test_legendre.py::test_orthogonal[tetrahedron] 	PASSED                                                                  [ 91%]
> test_legendre.py::test_orthogonal[quadrilateral] PASSED                                                                [ 93%]
> test_legendre.py::test_orthogonal[hexahedron] PASSED                                                                   [ 95%]
> test_legendre.py::test_orthogonal[prism] PASSED                                                                        [ 97%]
> test_legendre.py::test_orthogonal[pyramid] SKIPPED (<Skipped instance>)                                                [100%]
> ```

I'm yet to implement Legendre polynomials on pyramid, so these are skipped for now (although perhaps xfail would be a better thing to do).
