### 2021-08-10 #3556 @mscroggs
> The software is well designed, using the general Ciarlet definition (`CiarletElement`) of the finite element as the starting point for most (all?) element types,

All except one (direct serendipity) so far are defined using the Ciarlet definition
