+++ joss/paper.bib
@@ -42,7 +42,6 @@
 @book{ciarlet,
   title = {The Finite Element Method for Elliptic Problems},
   author = {Philippe G. Ciarlet},
-  year = {1978},
-
+  year = {2002},
   doi = {10.1137/1.9780898719208}
 }
