### 2021-08-10 cab4eca
add some examples to readme
