### 2021-08-10 #3556 @mscroggs
I've added an extra demo, added more information to the README, and adjusted what Read the Docs says in mscroggs/symfem#118. (although the changes won't appear on Read the Docs until I next release a version)
