+++ demo/nedelec.py
@@ -1,5 +1,6 @@
 """
-This demo shows how Symfem can be used to verify properties of the polynomial basis of an element.
+This demo shows how Symfem can be used to verify properties of the polynomial basis
+and basis functions of an element.
 
 The polynomial set of a degree k Nedelec first kind space is:
 {polynomials of degree < k} UNION {polynomials of degree k such that p DOT x = 0}.
