+++ demo/lagrange.py
@@ -24,22 +24,19 @@ basis = element.get_basis_functions()
 edge_basis = [subs(f, x[0], a) for f in edge_element.get_basis_functions()]
 
 # Get the DOFs on edge 0 (from vertex 1 (1,0) to vertex 2 (0,1))
+# (1 - a, a) is a parametrisation of this edge
 dofs = element.entity_dofs(0, 1) + element.entity_dofs(0, 2) + element.entity_dofs(1, 0)
 # Check that the basis functions on this edge are equal
-# (1 - a, a) is a parametrisation of this edge
 for d, edge_f in zip(dofs, edge_basis):
+    # symequal will simplify the expressions then check that they are equal
     assert symequal(subs(basis[d], x[:2], (1 - a, a)),  edge_f)
 
-# Get the DOFs on edge 1 (from vertex 0 (0,0) to vertex 2 (0,1))
+# Get the DOFs on edge 1 (from vertex 0 (0,0) to vertex 2 (0,1), parametrised (0, a))
 dofs = element.entity_dofs(0, 0) + element.entity_dofs(0, 2) + element.entity_dofs(1, 1)
-# Check that the basis functions on this edge are equal
-# (0, a) is a parametrisation of this edge
 for d, edge_f in zip(dofs, edge_basis):
     assert symequal(subs(basis[d], x[:2], (0, a)),  edge_f)
 
-# Get the DOFs on edge 2 (from vertex 0 (0,0) to vertex 1 (1,0))
+# Get the DOFs on edge 2 (from vertex 0 (0,0) to vertex 1 (1,0), parametrised (a, 0))
 dofs = element.entity_dofs(0, 0) + element.entity_dofs(0, 1) + element.entity_dofs(1, 2)
-# Check that the basis functions on this edge are equal
-# (a, 0) is a parametrisation of this edge
 for d, edge_f in zip(dofs, edge_basis):
     assert symequal(subs(basis[d], x[:2], (a, 0)),  edge_f)
