+++ docs/demos/index.rst
@@ -5,6 +5,7 @@ Symfem demos
 .. toctree::
    :titlesonly:
 
+   lagrange
+   nedelec
    stiffness_matrix
    custom_element
-
