### 2021-08-11 dcb34fd
Corrected mapping in Lagrange demo
