### 2021-08-11 #3556 @sigvaldm
I think the new demos on the Lagrange and Nedelec elements are perfect examples of what this software can do, that other software can't. I got an `AssertionError` when running the `lagrange.py`, though, but I think I could understand the intention, and was able to make it pass. See if you're happy with the PR, or if not, make sure `lagrange.py` doesn't fail. There was also another error, but I didn't dig into that:

```
[sigvald@sim demo]$ pytest -v .
==================================================================================================================== test session starts =====================================================================================================================
platform linux -- Python 3.9.5, pytest-6.2.1, py-1.10.0, pluggy-0.13.1 -- /usr/bin/python
cachedir: .pytest_cache
rootdir: /home/sigvald/Codes/symfem
plugins: cov-2.11.1
collected 5 items

test_demos.py::test_demo[lagrange.py] FAILED                                                                                                                                                                                                           [ 20%]
test_demos.py::test_demo[sim.py] PASSED                                                                                                                                                                                                                [ 40%]
test_demos.py::test_demo[nedelec.py] PASSED                                                                                                                                                                                                            [ 60%]
test_demos.py::test_demo[stiffness_matrix.py] PASSED                                                                                                                                                                                                   [ 80%]
test_demos.py::test_demo[custom_element.py] FAILED
```
