### 2021-08-11 a7f93c9
Add fast test running to readme
