### 2021-08-11 #3556 @sigvaldm
> I've added a reference in mscroggs/symfem@1398dfd

Looks good. Once it's merged into the main branch I'd be happy to tick off the "References" and "Quality of writing" checkboxes.
