### 2021-08-11 #3556 @sigvaldm
> I'm yet to implement Legendre polynomials on pyramid, so these are skipped for now (although perhaps xfail would be a better thing to do).

Now that I know the reason I think it's fine (you don't claim it to be implemented anywhere), and I don't have a strong opinion on whether to use `skip` or `xfail`. Using e.g. `reason="not yet implemented"` would make it obvious why it is skipped, though.
