### 2021-08-11 fd46ca8
Update docs (#118)
* Add reference to Ciarlet

* DOI is for 2002 reprint

* add some examples to readme

* Add Nedelec demo
Add text to demos
Test snippets in readme

* Add a LAgrange test

* update demos

* Add demos to docs

* add more intro to readme

* Corrected mapping in Lagrange demo (#119)

Co-authored-by: Matthew Scroggs <matthew.w.scroggs@gmail.com>

* Add fast test running to readme

* give reason for xfailing

Co-authored-by: Sigvald Marholm <sigvald@marebakken.com>
