+++ joss/paper.bib
@@ -38,3 +38,10 @@
   doi = {10.11588/ans.2015.100.20553},
   pages = {9--23},
 }
+
+@book{ciarlet,
+  title = {The Finite Element Method for Elliptic Problems},
+  author = {Philippe G. Ciarlet},
+  year = {2002},
+  doi = {10.1137/1.9780898719208}
+}
