+++ test/test_against_computed_by_hand.py
@@ -1,11 +1,11 @@
 import sympy
 from symfem import create_element
-from symfem.symbolic import x, all_symequal
+from symfem.symbolic import x, symequal
 
 
 def test_lagrange():
     space = create_element("triangle", "Lagrange", 1)
-    assert all_symequal(
+    assert symequal(
         space.tabulate_basis([[0, 0], [0, 1], [1, 0]]),
         ((1, 0, 0), (0, 0, 1), (0, 1, 0)),
     )
@@ -13,7 +13,7 @@ def test_lagrange():
 
 def test_nedelec():
     space = create_element("triangle", "Nedelec", 1)
-    assert all_symequal(
+    assert symequal(
         space.tabulate_basis([[0, 0], [1, 0], [0, 1]], "xxyyzz"),
         ((0, 0, 1, 0, 1, 0), (0, 0, 1, 1, 0, 1), (-1, 1, 0, 0, 1, 0)),
     )
@@ -21,7 +21,7 @@ def test_nedelec():
 
 def test_rt():
     space = create_element("triangle", "Raviart-Thomas", 1)
-    assert all_symequal(
+    assert symequal(
         space.tabulate_basis([[0, 0], [1, 0], [0, 1]], "xxyyzz"),
         ((0, -1, 0, 0, 0, 1), (-1, 0, -1, 0, 0, 1), (0, -1, 0, -1, 1, 0)),
     )
@@ -29,7 +29,7 @@ def test_rt():
 
 def test_Q():
     space = create_element("quadrilateral", "Q", 1)
-    assert all_symequal(
+    assert symequal(
         space.tabulate_basis([[0, 0], [1, 0], [0, 1], [1, 1]]),
         ((1, 0, 0, 0), (0, 1, 0, 0), (0, 0, 1, 0), (0, 0, 0, 1)),
     )
@@ -38,7 +38,7 @@ def test_Q():
 def test_dual0():
     space = create_element("dual polygon(4)", "dual", 0)
     q = sympy.Rational(1, 4)
-    assert all_symequal(
+    assert symequal(
         space.tabulate_basis([[q, q], [-q, q], [-q, -q], [q, -q]]),
         ((1, ), (1, ), (1, ), (1, ))
     )
@@ -49,7 +49,7 @@ def test_dual1():
     h = sympy.Rational(1, 2)
     q = sympy.Rational(1, 4)
     e = sympy.Rational(1, 8)
-    assert all_symequal(
+    assert symequal(
         space.tabulate_basis([[0, 0], [q, q], [h, 0]]),
         ((q, q, q, q),
          (sympy.Rational(5, 8), e, e, e),
@@ -67,11 +67,11 @@ def test_lagrange_pyramid():
              (1 - x_i) * y_i / (1 + z_i),
              x_i * y_i / (1 + z_i),
              z_i / (1 + z_i)]
-    assert all_symequal(basis, space.get_basis_functions())
+    assert symequal(basis, space.get_basis_functions())
 
     basis = [(1 - x[0] - x[2]) * (1 - x[1] - x[2]) / (1 - x[2]),
              x[0] * (1 - x[1] - x[2]) / (1 - x[2]),
              (1 - x[0] - x[2]) * x[1] / (1 - x[2]),
              x[0] * x[1] / (1 - x[2]),
              x[2]]
-    assert all_symequal(basis, space.get_basis_functions())
+    assert symequal(basis, space.get_basis_functions())
