### 2021-08-11 cbeedc6
Add CHANGELOG.md and use CHANGELOG_SINCE_LAST_VERSION.md to set release description
