### 2021-08-11 68de155
Add changelog (#120)
* Add CHANGELOG.md and use CHANGELOG_SINCE_LAST_VERSION.md to set release description

* ++version

* put date in changelog
