+++ .github/scripts/version.py
@@ -1,6 +1,7 @@
 import json
 import sys
 import github
+from datetime import datetime
 
 access_key = sys.argv[-1]
 
@@ -22,6 +23,24 @@ for release in symfem.get_releases():
     if release.tag_name == f"v{version}":
         break
 else:
+    changelog_file = symfem.get_contents("CHANGELOG_SINCE_LAST_VERSION.md", branch.commit.sha)
+    changes = changelog_file.decoded_content.decode("utf8").strip()
+    if changes == "":
+        changes = "- Released new version of Symfem"
     symfem.create_git_tag_and_release(
-        f"v{version}", f"Version {version}", f"Version {version}", "Latest release",
+        f"v{version}", f"Version {version}", f"Version {version}", changes,
         branch.commit.sha, "commit")
+
+    old_changelog_file = symfem.get_contents("CHANGELOG.md", branch.commit.sha)
+    old_changes = old_changelog_file.decoded_content.decode("utf8").strip()
+
+    new_changelog = (f"# Version {version} ({datetime.now().strftime('%d %B %Y')})\n\n"
+                     f"{changes}\n\n{old_changes}")
+
+    symfem.update_file(
+        "CHANGELOG.md", "Update CHANGELOG.md", new_changelog, sha=old_changelog_file.sha
+    )
+    symfem.update_file(
+        "CHANGELOG_SINCE_LAST_VERSION.md", "Reset CHANGELOG_SINCE_LAST_VERSION.md", "",
+        sha=changelog_file.sha
+    )
