+++ CHANGELOG.md
@@ -1,0 +1,3 @@
+# 11 August 2021
+
+- Started keeping changelog
