+++ .github/scripts/version.py
@@ -21,26 +21,29 @@ assert data["version"] == version
 
 for release in symfem.get_releases():
     if release.tag_name == f"v{version}":
-        break
+        pass  # break
 else:
-    changelog_file = symfem.get_contents("CHANGELOG_SINCE_LAST_VERSION.md", branch.commit.sha)
+    new_branch = symfem.create_git_ref(ref=f"refs/heads/v{version}-changelog", sha=branch.commit.sha)
+    changelog_file = symfem.get_contents("CHANGELOG_SINCE_LAST_VERSION.md", new_branch.commit.sha)
     changes = changelog_file.decoded_content.decode("utf8").strip()
-    if changes == "":
-        changes = "- Released new version of Symfem"
-    symfem.create_git_tag_and_release(
-        f"v{version}", f"Version {version}", f"Version {version}", changes,
-        branch.commit.sha, "commit")
+    # if changes == "":
+    #    changes = "- Released new version of Symfem"
+    # symfem.create_git_tag_and_release(
+    #    f"v{version}", f"Version {version}", f"Version {version}", changes,
+    #    branch.commit.sha, "commit")
 
-    old_changelog_file = symfem.get_contents("CHANGELOG.md", branch.commit.sha)
-    old_changes = old_changelog_file.decoded_content.decode("utf8").strip()
+    if changes != "":
 
-    new_changelog = (f"# Version {version} ({datetime.now().strftime('%d %B %Y')})\n\n"
-                     f"{changes}\n\n{old_changes}")
+        old_changelog_file = symfem.get_contents("CHANGELOG.md", new_branch.commit.sha)
+        old_changes = old_changelog_file.decoded_content.decode("utf8").strip()
 
-    symfem.update_file(
-        "CHANGELOG.md", "Update CHANGELOG.md", new_changelog, sha=old_changelog_file.sha
-    )
-    symfem.update_file(
-        "CHANGELOG_SINCE_LAST_VERSION.md", "Reset CHANGELOG_SINCE_LAST_VERSION.md", "",
-        sha=changelog_file.sha
-    )
+        new_changelog = (f"# Version {version} ({datetime.now().strftime('%d %B %Y')})\n\n"
+                         f"{changes}\n\n{old_changes}")
+
+        symfem.update_file(
+            "CHANGELOG.md", "Update CHANGELOG.md", new_changelog, sha=old_changelog_file.sha
+        )
+        symfem.update_file(
+            "CHANGELOG_SINCE_LAST_VERSION.md", "Reset CHANGELOG_SINCE_LAST_VERSION.md", "",
+            sha=changelog_file.sha
+        )
