+++ .github/scripts/version.py
@@ -23,7 +23,8 @@ for release in symfem.get_releases():
     if release.tag_name == f"v{version}":
         pass  # break
 else:
-    new_branch = symfem.create_git_ref(ref=f"refs/heads/v{version}-changelog", sha=branch.commit.sha)
+    symfem.create_git_ref(ref=f"refs/heads/v{version}-changelog", sha=branch.commit.sha)
+    new_branch = symfem.get_branch(f"v{version}-changelog")
     changelog_file = symfem.get_contents("CHANGELOG_SINCE_LAST_VERSION.md", new_branch.commit.sha)
     changes = changelog_file.decoded_content.decode("utf8").strip()
     # if changes == "":
@@ -47,3 +48,8 @@ else:
             "CHANGELOG_SINCE_LAST_VERSION.md", "Reset CHANGELOG_SINCE_LAST_VERSION.md", "",
             sha=changelog_file.sha
         )
+
+        symfem.create_pull(
+            title="Update changelogs", body="", base="main", head=f"v{version}-changelog"
+        )
+
