+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+# Version 2021.8.1 (11 August 2021)
+
+- Added Guzman-Neilan element
+- Added nonconforming Arnold-Winther element
+- Wrote JOSS paper
+
 # 11 August 2021
 
 - Started keeping changelog
