+++ CHANGELOG_SINCE_LAST_VERSION.md
@@ -1,3 +1,0 @@
-- Added Guzman-Neilan element
-- Added nonconforming Arnold-Winther element
-- Wrote JOSS paper
