+++ .github/scripts/version.py
@@ -21,36 +21,33 @@ assert data["version"] == version
 
 for release in symfem.get_releases():
     if release.tag_name == f"v{version}":
-        pass  # break
+        break
 else:
     symfem.create_git_ref(ref=f"refs/heads/v{version}-changelog", sha=branch.commit.sha)
     new_branch = symfem.get_branch(f"v{version}-changelog")
     changelog_file = symfem.get_contents("CHANGELOG_SINCE_LAST_VERSION.md", new_branch.commit.sha)
     changes = changelog_file.decoded_content.decode("utf8").strip()
-    # if changes == "":
-    #    changes = "- Released new version of Symfem"
-    # symfem.create_git_tag_and_release(
-    #    f"v{version}", f"Version {version}", f"Version {version}", changes,
-    #    branch.commit.sha, "commit")
+    if changes == "":
+        changes = "- Released new version of Symfem"
+    symfem.create_git_tag_and_release(
+        f"v{version}", f"Version {version}", f"Version {version}", changes,
+        branch.commit.sha, "commit")
 
-    if changes != "":
+    old_changelog_file = symfem.get_contents("CHANGELOG.md", new_branch.commit.sha)
+    old_changes = old_changelog_file.decoded_content.decode("utf8").strip()
 
-        old_changelog_file = symfem.get_contents("CHANGELOG.md", new_branch.commit.sha)
-        old_changes = old_changelog_file.decoded_content.decode("utf8").strip()
+    new_changelog = (f"# Version {version} ({datetime.now().strftime('%d %B %Y')})\n\n"
+                     f"{changes}\n\n{old_changes}\n")
 
-        new_changelog = (f"# Version {version} ({datetime.now().strftime('%d %B %Y')})\n\n"
-                         f"{changes}\n\n{old_changes}\n")
+    symfem.update_file(
+        "CHANGELOG.md", "Update CHANGELOG.md", new_changelog, sha=old_changelog_file.sha,
+        branch=f"v{version}-changelog"
+    )
+    symfem.update_file(
+        "CHANGELOG_SINCE_LAST_VERSION.md", "Reset CHANGELOG_SINCE_LAST_VERSION.md", "",
+        sha=changelog_file.sha, branch=f"v{version}-changelog"
+    )
 
-        symfem.update_file(
-            "CHANGELOG.md", "Update CHANGELOG.md", new_changelog, sha=old_changelog_file.sha,
-            branch=f"v{version}-changelog"
-        )
-        symfem.update_file(
-            "CHANGELOG_SINCE_LAST_VERSION.md", "Reset CHANGELOG_SINCE_LAST_VERSION.md", "",
-            sha=changelog_file.sha, branch=f"v{version}-changelog"
-        )
-
-        symfem.create_pull(
-            title="Update changelogs", body="", base="main", head=f"v{version}-changelog"
-        )
-
+    symfem.create_pull(
+        title="Update changelogs", body="", base="main", head=f"v{version}-changelog"
+    )
