### 2021-08-11 fab895b
update link style in docs index
