+++ docs/index.rst
@@ -203,4 +203,4 @@ Documentation index
    demos/index
    docs/index
 
-.. [Ciarlet] P. G. Ciarlet, The Finite Element Method for Elliptic Problems (2002, first published 1978) [DOI: 10.1137/1.9780898719208](https://doi.org/10.1137/1.9780898719208)
+.. [Ciarlet] Ciarlet, P. G., The Finite Element Method for Elliptic Problems (2002, first published 1978) `DOI: 10.1137/1.9780898719208 <https://doi.org/10.1137/1.9780898719208>`_
