### 2021-08-11 #3556 @mscroggs
> > I've added a reference in [mscroggs/symfem@1398dfd](https://github.com/mscroggs/symfem/commit/1398dfd)
>
> Looks good. Once it's merged into the main branch I'd be happy to tick off the "References" and "Quality of writing" checkboxes.

It's merged. I've also bumped the version number so that the Read the Docs documentation is now up to date
