### 2021-08-11 #3556 @sigvaldm
The documentation has been enhanced in several places. Especially the two extra demos are enlightening, as they demonstrate the capabilities unique to this software. I recommend that this submission be accepted (@Nikoleta-v3).
