### 2021-08-16 7b35b4b
Link to changelog in readme. Check that CHANGELOG_SINCE_LAST_VERSION is not empty when bumping version number
