### 2021-08-16 1b9037a
Don't release version if CHANGELOG_SINCE_LAST_VERSION.md is empty
