### 2021-08-20 #3556 @iammix
@mscroggs Great work!
The structure of the code is clean and clear. I like that you added the corresponding papers in docstrings of each element implementation.
Regarding the tests, it is good that you used **xfail** for the not-implemented features. It advises beforehand the user what will follow.
I would recommend you to numerate the equation in the paper. After that, the paper is ready for acceptance from my side.

@Nikoleta-v3 , after the aforementioned change, the submission can be accepted.🎉
