### 2021-08-20 e854224
equation not $$ (#122)
