### 2021-08-22 #3556 @jmv2009
I don't know if can throw in my 2 cents here: Discussion about advantages of difference methodology see e.g. sagemath usage in https://arxiv.org/pdf/1806.00031.pdf. Also refs for individual geometries are not always clear (e.g. prism nedelec1) https://defelement.com/elements/nedelec1.html (And not following https://arxiv.org/abs/1605.00132)
