### 2021-08-22 #3556 @mscroggs
@jmv2009 Sorry, I don't follow your comment about difference methodology.

Re refs: This looks like an issue related to the information on DefElement, not Symfem. I've opened some issues on the DefElement repo (https://github.com/mscroggs/defelement.com) about making the references there more complete. If you spot any further issues in the DefElement references, you can open more issues there and I'll fix them as soon as I have time.
