+++ .github/workflows/style-checks.yml
@@ -2,6 +2,11 @@ name: 🕶️ Style
 
 on:
   push:
+    branches:
+      - "**"
+  pull_request:
+    branches:
+      - main
 
 jobs:
   style-checks:
