+++ .github/workflows/coveralls.yml
@@ -37,7 +37,7 @@ jobs:
         name: Run unit tests
         continue-on-error: true
       - name: Upload to Coveralls
+        if: ${{ github.ref == 'ref/head/main' }}
         env:
           COVERALLS_REPO_TOKEN: ${{ secrets.coverall_token }}
         run: coveralls
-        if: {{ github.ref == 'ref/head/main' }}
