### 2021-08-23 5fbc5b3
run all the tests
