+++ .github/scripts/coveralls_check.py
@@ -1,0 +1,2 @@
+with open(".coveralls_output") as f:
+    assert int(f.read().split("TOTAL")[-1].split("%")[0].split(" ")[-1]) > 80
