### 2021-08-23 50a2f52
Fix CI on PR from fork (#125)
* Run style checks on PR

* speed up tests for testing

* only upload to coveralls from main

* $

* assert coveralls is hgh enough

* run all the tests

* \?

* Coveralls check

* ignore coveralls

* .github/
