### 2021-08-24 #3556 @Nikoleta-v3
No further comments. Everything looks good to me 👍🏻

At this point could you:

The tagged release is `v2021.7.6`? (sanity check) 😅
