### 2021-08-24 4e2acaf
Changelog, add conda to read the docs
