+++ CHANGELOG_SINCE_LAST_VERSION.md
@@ -1,0 +1,1 @@
+- New release for JOSS paper
