+++ CITATION.cff
@@ -5,6 +5,6 @@ authors:
     given-names: Matthew W.
     orcid: 0000-0002-4658-2443
 title: Symfem
-version: 2021.8.1
-date-released: 2021-08-11
+version: 2021.8.2
+date-released: 2021-08-24
 license: MIT
