+++ .github/workflows/test-packages.yml
@@ -18,7 +18,7 @@ jobs:
           python-version: ${{ matrix.python-version }}
       - uses: actions/checkout@v2
         with:
-          ref: v2021.8.1
+          ref: v2021.8.2
         name: Get latest Symfem version
       - run: rm -rf symfem VERSION
         name: Remove downloaded symfem
@@ -45,7 +45,7 @@ jobs:
           activate-environment: symfem
       - uses: actions/checkout@v2
         with:
-          ref: v2021.8.1
+          ref: v2021.8.2
         name: Get latest Symfem version
       - run: rm -rf symfem VERSION
         name: Remove downloaded symfem
@@ -71,13 +71,13 @@ jobs:
         with:
           python-version: ${{ matrix.python-version }}
       - run: |
-          wget -O symfem.tar.gz https://pypi.io/packages/source/s/symfem/symfem-2021.8.1.tar.gz
+          wget -O symfem.tar.gz https://pypi.io/packages/source/s/symfem/symfem-2021.8.2.tar.gz
           tar -xvzf symfem.tar.gz
         name: Download and unpack latest version of Symfem
       - run: pip install pytest
         name: Install pytest
       - run: |
-          cd symfem-2021.8.1
+          cd symfem-2021.8.2
           pip install -r requirements.txt
           python3 -m pytest test/
         name: Install requirements, run tests and demos
