+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+# Version 2021.8.2 (24 August 2021)
+
+- New release for JOSS paper
+
 # Version 2021.8.1 (11 August 2021)
 
 - Added Guzman-Neilan element
