+++ CHANGELOG_SINCE_LAST_VERSION.md
@@ -1,1 +1,0 @@
-- New release for JOSS paper
