### 2021-08-24 #3556 @mscroggs
> I see you listed twice in the zenodo archive metadata, which looks odd and which you can fix - I'll continue processing the paper, as this doesn't really affect the process, but would good to clear up before or soon after we finish.
@danielskatz Thanks, I've fixed that - I'd not realised that authors and contributors were displayed next to each other
